requirements.txt
gunicorn
fastapi
uvicorn
annotated-types==0.7.0
anyio==4.9.0
bcrypt==3.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emails==0.6
fastapi==0.109.2
greenlet==3.2.1
h11==0.16.0
idna==3.10
Jinja2==3.1.2
lxml==5.4.0
MarkupSafe==3.0.2
more-itertools==10.7.0
packaging==25.0
passlib==1.7.4
pipdeptree==2.26.1
premailer==3.10.0
pyasn1==0.6.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.36.3
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.27.1
